move product-solver higher up
o move product-solver for equations that are explicitly products higher up in the solve process.

o simplified scheme for determining if an expression is a kernel.

o for a term XXX^n with integer n, recogonize XXX, not XXX^n as the kernel.

o Append to rtest_extra_solve

o 97 tests failed out of 16,898 total tests.
Evaluation took:
  417.507 seconds of real time
  407.328125 seconds of total run time (379.875000 user, 27.453125 system)
  [ Run times consist of 20.131 seconds GC time, and 387.198 seconds non-GC time. ]
  97.56% CPU
  446,502,859 forms interpreted
  1,378 lambdas converted
  1,132,277,218,600 processor cycles
  134,510,868,080 bytes consed